[SDESK-1374] - Filter condition field name improvements
