+++ scripts/apps/content-filters/views/manage-filter-conditions.html
@@ -43,7 +43,7 @@
                 <div class="field">
                     <label translate>Field</label>
                     <select class="input" required ng-model="filterCondition.field" ng-change="clearConditionValues()">
-                        <option value="{{parameter.field}}" ng-selected="parameter.field === filterCondition.field" ng-repeat="parameter in filterConditionParameters | orderBy: 'field' track by $index">{{:: parameter.field}}</option>
+                        <option value="{{parameter.field}}" ng-selected="parameter.field === filterCondition.field" ng-repeat="parameter in filterConditionParameters | orderBy: 'field' track by $index">{{:: label(parameter.field) }}</option>
                     </select>
                 </div>
                 <div class="field">
